+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added pagination to search results.
+- Display emoji reactions under post.
+- Added picker for emoji reactions.
 
 ## [3.1.0] - 2024-08-08